Show correct diffs for conflicts while rebasing
This patch addresses issue reported in #407 by @qstrahl. Recent versions
of git have changed the way statuses are displayed for unmerged files
while rebasing. The result is that unmerged files in a rebase conflict
are treated like staged files. This patch corrects this issue by
re-ordering the checks that are performed to determine which status
section the file is located in so that the check for the unmerged file
happens before the check for the staged file.