Detect 'git worktree add' checkouts
The worktree feature of 2.5.0 stores a slimmed down .git directory under
the real .git directory:

    .git/worktree/$name

Under this directory, the files we care about:

    commondir -> relative path to the common .git directory
    gitdir -> absolute path to the .git file in the new worktree
    gitfile -> relative path to the .git file from the main worktree

The objects/, refs/, and other .git directories are contained in the
common .git directory. Code needs to look in the right place for things
depending on what is queried.

Fixes #655.