Do not call fnameescape globally in EditComplete
EditComplete is the only part which fnameescapes the whole value, which
is problematic for refs like 'stash@{0}', which becomes 'stash@\{0}'.
Only the filename part should be escaped, and that gets done in
repo.superglob now.

It would be better to not include the NUL as <C-V><C-A>, but I have not
managed to do so: trying to split on `[\x0]` seems to make `split`
consider the pattern empty and then use spaces instead?!