Replace.cmd: make sure, the window stays the same
If you view a large patch file from git, (e.g. from :Gblame), the large
file plugin might be triggered, which runs the NoMatchParen command,
which internally uses windo to iterate over all windows to disable match
parent highlighting.

However once the command has been run, we end up in the wrong window.
Now this is actually a bug in the match paren plugin or large file
plugin. But in any case add a check, that this does not cause a problem.