Provide some commit --reset-author maps
Supporting flags with additional maps is a pretty limited solution, as
there's no way to combine them (that is, we'd need to map both cRva and
cvRa to get both verbose and reset-author, and that's as silly as it is
hard to type in under a second).  I think some kind of getchar()
dispatcher is going to be the long term solution.