Fix script number in s:function()
Problem:

Assume the following code in your vimrc:

    nnoremap <leader>n :call <sid>foo()<cr>

    function! s:foo()
      Gllog
    endfunction

Using that mapping would result in:

    E700: Unknown function: <SNR>1_LogParse

This happens because expand('<sfile>') in s:function() evaluates to:

    <SNR>1_foo[1]..fugitive#LogCommand[74]..<SNR>68_function

Then the matchstr() would extract the wrong `<SNR>1_` (vimrc) instead of
the correct `<SNR>68_` (autoload/fugitive.vim).

Solution:

Adjust the regexp to extract the script number of the last function in
the call chain, s:function() itself, which is guaranteed to be in the
correct script, autoload/fugitive.vim.