Include git notes in commit buffers
Issue #912

Add any existing git notes into commit views, between the commit message
and the patch itself.  This is similar to what "git show" does, albeit
without any indent.

This should work for older versions of git, without the notes feature,
provided that the '%N' directive is passed through as-is.