Hash colors for :Gblame using 256 color mode in terminal
Using only colors from the 6x6x6 cube without system colors and the
grayscale ramp at the end of the 256 color mode.

Those colors also have the nice property of not being too dark or too
bright and therefore not making the hash unreadable on a dark or light
background.