Provide default merge head for "." remote
Almost included this in f75dbb19672b590fe8e32334051314b47e91ae02, but
dropped it because it always equal to the current branch, guaranteeing
an empty log, and because unlike the "origin" default, it shouldn't
really happen in practice.  However if a user screws up their config in
a way that forces it to happen, we should still try to avoid the same
"Push" header issue.